Don't overly create Python-object based ASTs
Summary:
Adds some optimizations so that we're not overly creating ASTs.  When we go through the `loader.check` API we end up creating the Python AST objects to return back when the module is strict.  This causes us to skip the API in two cases:

1) We already have the module loaded
2) We've already attempted to load the module, but it's not static

This greatly reduces the load time when doing lots of imports from source and should fix the issues w/ the test runners not starting up with strictified `util.etc`.

I think the problem w/ `util.etc` is it probably increases the time a little bit being a new strict module which is commonly imported, and testing it against a non-bolted pre-production runtime greatly pushes the time up.  This causes the tests to now hit the timeout.

But I'm also not sure that this was any kind of regression, it seems like we've just been slowing creeping up over the past month: https://fburl.com/scuba/ig_srv_test_actions/eft64ejc

Reviewed By: czardoz

Differential Revision: D32123646

fbshipit-source-id: fd3c3140d8