simplify classmethod flag to INVOKE_METHOD
Summary:
In updating how property type descrs and thunks work, I was bothered by the false
parallel between the type-descr format for properties and classmethods. Both currently
use a similar format `("module", "Class", ("attr", "marker"))`, but for a property
the marker is an actual attribute of the property object, whereas in the classmethod case
it's an arbitrary marker that is ignored by the classloader itself, it's just an extra flag to
the interpreter / jit compiler about how to treat the top-of-stack (as type or instance.)
Given this, it seems more logical (and simplifies the code quite a bit) if we don't try to
integrate this flag into the type-descr at all, but rather make it an optional extra flag directly
to the `INVOKE_METHOD` opcode.

The only downside I can see is that it makes the `INVOKE_METHOD`
oparg format slightly less self-documenting, but I think this is arguable: `("foo", "__class__")`
isn't very clear either until you learn it. IMO the code simplification of not
having to handle the alternate form of type-descr throughout the classloader is worth it.

Reviewed By: DinoV

Differential Revision: D32118885

fbshipit-source-id: bd3b67281a