rework property handling in vtables
Summary:
This fixes some edge cases in property patching.

It ensures that in `vt_slotmap` and `vt_originals` we will always have keys for the property itself
and for its getter and setter (under `("propname", "fget")` and `("propname", "fset")` keys),
and in `vt_thunks` we'll have the getter and setter keys. And then we ensure that all of these,
and the vtable entries themselves, are updated appropriately on patching.

In order to reproduce the correct `AttributeError` for a missing getter/setter, and avoid cascading
complexity with handling the `NULL` values for them, we add builtin stub functions in `_static`
for this behavior, and return those on lookup rather than `NULL`.

This also requires adding some handling for builtin typed methods in the classloader. For
now I only support resolving return types of `_Py_SIG_VOID` and `_Py_SIG_ERROR`; the
latter is the only one needed in this diff and the former has the same return type; we can
add support for others as we need them.

The signature of `get_func_or_special_callable` (and thus also
`_PyClassLoader_GetStaticallyInheritedMember`) needed to change to return an error code
and the lookup result as an out param, because there was ambiguity in whether a `NULL`
return meant "not found in dict" (which is not necessarily an error) vs "an error occurred"
(the only error currently is finding a non-property in a slot where a property is expected.)

Reviewed By: sinancepel

Differential Revision: D32160089

fbshipit-source-id: 37f41343c9