Add temporary RefineTypes to truthy blocks of CondBranchCheckType
Summary:
If we emit a `CondBranchCheckType` to a `BasicBlock` with only a single
edge in, we can refine the type of the target to the checked `Type` in
the truthy block.

Since `RefineType` doesn't get translated down to LIR, this should just
be a no-op which provides us with more static type information.

Reviewed By: carljm

Differential Revision: D32178877

fbshipit-source-id: 5eae9524eb