Add and expose _PyCheckedList_TypeCheck and _PyCheckedDict_TypeCheck
Summary:
This adds a way for the JIT to determine whether a `PyTypeObject` is
a generic type instantiation of `chklist` or `chkdict`.

This involves adding a more general
`_PyClassLoader_GetGenericTypeDefFromType` function in order to get the
generic type definition of a type, and not just of instances of a type.

Reviewed By: carljm

Differential Revision: D32025263

fbshipit-source-id: 3abb42c4bb