Support patching properties with user defined descriptors
Summary:
Adds support for overriding and patching properties with user defined descriptors.  This primarily enables mocking out properties inside of tests.

We add a new thunk object that is used for dispatching to the descriptor methods of the underlying property.  On a get if the object isn't a descriptor we simply return the patched value.

The biggest thing to sort through here is that we need to dynamically create the thunks, so a bunch of functions which pass out values now have ref counting semantics where they return owned refs.  The thunk creation happens in `classloader_get_property_fget` and `classloader_get_property_fset` so this just propagates up through their callers.

Reviewed By: carljm

Differential Revision: D32197628

fbshipit-source-id: f9685ca980