Use PyObject_TypeCheck in type_call instead of IsSubtype equality test
Summary:
In D14351113, the equality check in IsSubtype was introduced as an optimization for the hot path of instance creation.

When upstreaming it ([bpo-45697](https://bugs.python.org/issue45697)) it was pointed out that it's preferable to use the `PyObject_TypeCheck` in the hot callsite
(which performs an inline equality check before calling to IsSubtype) instead of adding a test to all paths
(including those that already use `PyObject_TypeCheck`, hence would hit the equality check twice).

This diff ports the [merged upstream PR](https://github.com/python/cpython/pull/29392) to cinder.

Reviewed By: carljm

Differential Revision: D32175230

fbshipit-source-id: 93080c58d0