Fix annAssign validation with future annotation
Summary:
This fixes a difference between the python compiler and the builtin C compiler.
In an `AnnAssign` into an attribute, even when no RHS exist, the base of the attribute on LHS still need to be valid. The compiler generates bytecode to load the base, but the python compiler does not. See added test case for example.

Reviewed By: czardoz

Differential Revision: D32140959

fbshipit-source-id: 517346dafb