Fix fix optimization level and assert optimization
Summary:
This brings the behavior the same regarding asserts for the py compiler and the C compiler.
In the C compiler, assert statements are decided to be generated based on the optimization level in the code generator, but in the python compiler it's done in the ast optimizer. This seems to cause small inconsistencies when the statements are put in conditional blocks. This diff brings the py compiler to also make the optimization in the code generator itself.

This also fixes a bug in `CodeGenerator` where the optimization level is not passed correctly to child generators.

Reviewed By: czardoz

Differential Revision: D32227088

fbshipit-source-id: 6068831abe