remove legacy peephole optimizations in pycompile
Summary: This removes peephole optimizations that used to exist in the C compiler in 3.7 from the python compiler. Keeping these optimizations lead to slightly more efficient code in very edge cases, but is not quite worth it because most cases are already covered in the AST optimizer, and the discrepancy is a maintenance cost.

Reviewed By: czardoz

Differential Revision: D32214944

fbshipit-source-id: 82f5de6cc6