Add `_AwaitingFuture` and use it in `wait_for`
Summary:
This introduces a new future type, `_AwaitingFuture`, for
use in `asyncio.wait_for`. We want `asyncio.wait_for` to
be reliably cancellable, while also propagating the `awaiter`
onto the given awaitable. An `_AwaitingFuture` fits this bill.

An `_AwaitingFuture` waits for another future to complete.
It completes with `None` when the supplied future completes
and doesn't support `set_exception` or `set_result`. It can be
cancelled, but does not propagate cancellation onto the
awaited future. When directly `await`-ed, the `_AwaitingFuture`
will propagate the awaiter onto the linked future.

Reviewed By: vladima

Differential Revision: D32240291

fbshipit-source-id: 82a7adcb89