Make _PyClassLoader_ResolveReturnType the source of truth
Summary:
Currently we have some checks to see if a function is static or not in `type_vtable_setslot` but those checks don't align with what we can get a static return type from `_PyClassLoader_ResolveReturnType` for.  In particular if we have a class method then we won't see that it's statically declared even though it is.

This changes it so we simply try and resolve the return type on the value in hand, and fallback to looking up on the derived type only if that doesn't work.

Reviewed By: carljm

Differential Revision: D32293067

fbshipit-source-id: 21ae9ab589