begin consolidating classloader lookups in preload pass
Summary:
The JIT (and especially multi-threaded compile) assumes the invariant that we won't have Python code
execution mid-JIT-compile. But classloader lookups can trigger new imports, which is Python code
execution, and this is even more likely in the presence of lazy imports. In order to unblock rollout of
multi-threaded compile (which in turn unblocks larger jit list), we need to isolate classloader lookups
in a separate initial step of JIT compilation.

This diff is the first step. It introduces a `Preloader` class that runs over the bytecode before compilation,
performs any needed classloader lookups, and holds onto the results for later query during lowering.

To keep diffs smaller, this diff handles only type and field lookups, not functions or methods. That means
there is still some duplication of classloader resolution functions between `preload.cpp` and `builder.cpp`;
that will go away in later diffs when invoke target resolution is also migrated to the preloader.

The timing and location of when we call `preload()` may also need adjustment in later diffs in order to
actually fix the problem with multi-threaded compile; for now I'm just doing it early in `BuildHIR` while I
get the preloader fleshed out.

Reviewed By: swtaarrs

Differential Revision: D32302084

fbshipit-source-id: a10f84c59a