Abort if input to SSAify has Phis
Summary:
tekknolagi hit this while writing a test recently: `SSAify` does weird
things if there are `Phi`s in the incoming program. It tries to give each
operand to the `Phi` a value along each incoming edge, resulting in a lot of
`Nullptr` types that don't need to be there.

We could in theory make this work, but it's simpler to just disallow it and say
that non-SSA HIR can't use `Phi`s.

Reviewed By: tekknolagi

Differential Revision: D32389253

fbshipit-source-id: f081195833