Support (un)boxing enums outside the JIT
Summary:
This change gives enum types a real type descr so we can box, unbox, and type-check them.
We represent primitive enums in the eval loop as a bare integer,
so boxing them is equivalent to calling their type
and unboxing them is resolving their `value` attribute.

Static callers should get an unboxed enum and non-static callers should get a boxed enum.
With primitive integers, this representation can be dealt with in `RETURN_PRIMITIVE`, but for enums,
we need dedicated function entry points which can box the result for a non-static caller.

Add cases to `PRIMITIVE_BOX` and `PRIMITIVE_UNBOX` that handle enums,
and add shadow bytecodes for those cases. Unbox enums in `CHECK_ARGS`.
Convert `RETURN_PRIMITIVE` back to take an oparg instead of a type descr.

Reviewed By: DinoV

Differential Revision: D32178262

fbshipit-source-id: 47f6135231