Fix Type::operator| for TBottom and specialized types
Summary:
`TBottom | specialized_type` should evaluate to `specialized_type`,
since `TBottom` is a subtype of everything. Restructure `Type::operator|` to
allow for this case.

This exposed an existing issue with `specSubtype()`, where we didn't think
`TBottom` was a subtype of any specialized types, since it had `kSpecTop` as
its specialization kind. Fix this by adding `kSpecBottom` and using it where
appropriate.

Reviewed By: carljm

Differential Revision: D32336294

fbshipit-source-id: 14aa2b942a