ensure we don't allow re-declaring functions matching import names
Summary:
A pattern we've seen is the wild is that modules try to conditionally import elements from modules such as `uwsgi`
and fall back to local declarations. Previously, we'd allow the function declaration, and emit invokes to it. In the
case the module is defined, this is incorrect. Let's disallow re-declaring these, and instead force people to move conditional imports
to non-static modules.

We can re-visit this behavior if this ends up being a major adoption blocker.

Reviewed By: DinoV

Differential Revision: D32399563

fbshipit-source-id: 8e9743db57