Fix array subscripting behavior
Summary:
Currently, we crash in the JIT if the array index type cannot
be determined. E.g:
```
a[some_non_primitive] = 7
```
Leads us to emit
```
PRIMITIVE_LOAD_CONST 7
...
STORE_SUBSCR
```
But `STORE_SUBSCR` cannot handle primitives (and is not supposed to).
So, this diff updates our type inference to not infer RHS as a primitive
if the index type cannot be determined.

Reviewed By: carljm

Differential Revision: D32476077

fbshipit-source-id: aa65d6ecb1