don't TP_ALLOC nested classes
Summary:
We can't `TP_ALLOC` a class defined inside a function, because it isn't externally accessible.

There was a test that was supposed to ensure we didn't, but the test didn't verify anything
because `CALL_FUNCTION` is part of building a class; the test needs to verify absence of
`TP_ALLOC`, and should run the function too.

At some point I think we need a deeper overhaul of how we handle qualified names and type
descrs and nested scopes; right now we have disparate ad-hoc handling for nested functions
and nested classes. But for now ensuring that any class not defined at module level is
`DYNAMIC_TYPE` suffices.

Reviewed By: DinoV

Differential Revision: D32498343

fbshipit-source-id: 1a09c116a0