add HIR parser/printer support for LoadArg type
Summary:
`LoadArg` HIR instruction has a type associated with it; add support for printing/parsing
this type in textual HIR. For non-static Python the type of an arg is always `TObject`, so
to avoid noise in non-static HIR, the type is only printed if it is not `TObject`, and is
optional on parsing, defaulting to `TObject`.

Motivation here is for next diff in stack to be able to easily test correct type of a `LoadArg` in
an HIR test.

Reviewed By: mpage

Differential Revision: D32516898

fbshipit-source-id: d65dbde4ee