Add a way for HIR instructions to define their operand types
Summary:
Add the virtual function, `GetOperandType` to `Instr` to allow us to
call `GetOperandType` while iterating over all `Instr`s in a `BasicBlock`.

Add the function that's actually called in `InstrT`, and use CRTP in order
to add a top-level bounds check in `InstrT` while deferring the return
to each opcode's own implementation.

Adds `makeTypeVec` in order to have an easy way to add basic
type information to `INSTR_CLASS`, and statically initialize
a vector of the expected type information of each opcode.

Adds an `OperandType` struct in order to store whether an opcode
should be of a specific type, or be a subclass of one of multiple
types represented by a `Constraint`
 ---
This is in preperation for a `TypeVerification` pass which will make
sure that our HIR is "well-typed".

Reviewed By: swtaarrs

Differential Revision: D31836766

fbshipit-source-id: af1fd3963e