(un)box static enums in the JIT
Summary:
Introduce a primitive `TCEnum` type, with a type specialization,
used for boxing and unboxing static enums in the JIT.

Rather than use this new type to represent enums in HIR,
continue to use `TCInt64` to reflect their internal representation.

Ensure that enums are appropriately boxed when entering JIT-ted code
and unboxed when returning from JIT-ted code.
This includes boxing arguments when we deopt.

Reviewed By: DinoV

Differential Revision: D31209822

fbshipit-source-id: 9f472d1875