multithreaded compile preloads everything first
Summary:
Move the preloader creation out of `BuildHIR` and all the way up to `pyjit.cpp`, where
multithreaded compile preloads every function it is going to compile before it relaxes gilstate
checks and spawns worker threads for multi-threaded compilation. This means that we do
all of our classloader and globals work (which could trigger Python code execution) when we
are still single-threaded, avoiding the possibility of deadlocks.

The Preloader now holds onto the code object, globals, builtins, and optionally function object
that it has preloaded compilation data for. This means that a preloader alone has everything we
will need for compilation, which avoids having to pass around multiple related arguments
everywhere.

Because a Preloader has everything we need, the multi-threaded compile work queue can be
simplified; instead of carrying both function and code objects, it always carries code objects,
which we just use as indices into a global map of code object to preloader. The global map is
necessary for two reasons: firstly because for re-entrant compilation (from `emitInvokeFunction`)
we need to be able to find the right already-created preloader, and secondly because
`threaded_compile.h` can't operate on preloaders directly; since it is referenced in `Jit/log.h`
it has to have minimal dependencies in the JIT.

The preloaders are owned by the global map, which must keep them all alive until multithreaded
compilation is finished, since deallocating a preloader will touch PyObject refcounts and create
a race condition. We pass const refs to the preloaders down through to HIR building. In order
to keep a const ref to the preloader as `HIRBuilder` member, we make `buildHIR` a free
function that creates the `HIRBuilder` initialized with the preloader.

Reviewed By: swtaarrs

Differential Revision: D32435476

fbshipit-source-id: 617b320047