handle cell args in checkArgType
Summary:
The preloader API that takes a locals index and returns its type (used for setting
the type on `LoadArg` instructions) didn't account for negative locals indices in
`CHECK_ARGS` meaning "this arg is a cell."

In practice this doesn't really matter at the moment, because the type of a `LoadArg`
will immediately get subsumed into its cell, and we don't have generic types in JIT
HIR so we don't track the type of object contained in a cell, so it'll come out as `object`
on the other end anyway. But I noticed it since I need to add this support to
`CHECK_ARGS` handling in the preloader anyway for the next diff in stack, and we
may as well handle this correctly so far as we can.

Reviewed By: DinoV

Differential Revision: D32516899

fbshipit-source-id: a576f404c3