preload types for static arg type checks
Summary:
Preload the Python types for static python argument type checks and pass them through
the `hir::Function` to codegen. This eliminates the final classloader call in compilation --
all remaining calls to classloader (and globals loads) in `Jit/` are either in preload or at
runtime. This should eliminate the possibility of Python code execution during compilation
and of deadlock between compile lock and GIL during multi-threaded compilation.

We now have arg-checks info stored somewhat redundantly in a few different ways in
preloader and `hir::Function`: we have a `_PyTypedArgsInfo` object with primitive args
info only, which is used at runtime for unboxing at the call boundary from non-static,
we have `check_arg_pytypes` which is used to generate assembly for static arg type
checks at entry from non-static, we have dedicated bools for `has_primitive_args` and
`has_primitive_first_arg`, and we also have a mapping from locals index to HIR `Type`
that is stored only in preloader and used only in HIR building. There's probably some
consolidation that could happen in these representations, but it'll require some
refactoring of the code that uses this data. The duplication also doesn't hurt much.
In any case, the consolidation will be clearer if it happens in separate follow-up diffs.

Reviewed By: mpage

Differential Revision: D32548648

fbshipit-source-id: 0da9b3346f