handle globals/builtins changes during preloading
Summary:
If a global changes its value as a result of side effects from preloading, we still have time
to compile with the updated value instead of compiling with an already-obsolete value and
thus guaranteeing deopts.

Also handle crash-free the case where a globals or builtins dict becomes unwatchable as a
side effect of code executed during preloading (e.g. due to a globals load triggering a lazy
import.)

Reviewed By: mpage

Differential Revision: D32566498

fbshipit-source-id: 8834c48733