Fix a race condition during function entry in shadow-frame mode
Summary:
This was quite the bug to track down. Before we dive into the details, lets
go over some background about shadow-frame mode. We'll use the following terms
throughout the commit:

- **shadow stack** - The call stack of `_PyShadowFrame` objects beginning
  at `tstate->shadow_frame`.
- **pyframe stack** - The call stack of `PyFrameObject` objects beginning
  at `tstate->frame`.

When shadow-frame mode is active, we don't allocate PyFrameObjects when
entering a jit-compiled function. As a result, the pyframe stack will
contain gaps that correspond to jit-compiled functions. We only materialize
PyFrameObjects for jit-compiled functions when `PyEval_GetFrame` is called.
Once the shadow stack has been fully materialized, we expect the shadow stack and
pyframe stack to be identical. That is, there should be a 1-1 correspondence
between shadow frames and PyFrameObjects. Each jit-compiled function assumes
that there is a frame on the pyframe stack and will attempt to unlink one
when it returns.

For example, given a logical call stack (TOS last) of `f0 <- f1 <- f2 <- f3`,
where `f1` and `f3` are jit-compiled, the shadow stack and pyframe stack will
look like the following prior to materializing the shadow stack:

```
Shadow stack    Pyframe stack
------------    -------------
f0              f0
f1              f2
f2
f3
```

They should be identical after the shadow stack is materialized:

```
Shadow stack    Pyframe stack
------------    -------------
f0              f0
f1              f1
f2              f2
f3              f3
```

This fixes a race condition that can accidentally erase a materialized
shadow frame from the pyframe stack when calling an interpreted function
from a jit-compiled function. For example, if `f0` is jit-compiled but
the stack has been materialized, we'll have stacks that are identical
and look like:

```
Shadow stack    Pyframe stack
------------    -------------
f0              f0
```

If the race is triggered when `f0` calls `f1`, we erase the entry
for `f0` from the pyframe stack, resulting in stacks that look like:

```
Shadow stack    Pyframe stack
------------    -------------
f0              f1
f1
```

When `f0` returns, it expects to to unlink a Python frame from the
pyframe stack and will find none. Alternatively, it could incorrectly
unlink the Python frame for its caller.

To understand how this can happen, lets go over how `PyFrameObject`s
for interpreted functions are allocated and linked into the pyframe stack.
When a frame is allocated its `f_back` field is set to point to
`tstate->frame`. Shortly thereafter, `tstate->frame` is set to the
newly allocated frame by the interpreter loop. There is an implicit
assumption that `tstate->frame` will not change between when the
frame is allocated and when it is set to `tstate->frame`.  That
assumption is invalid when the JIT is executing in shadow-frame mode.

`tstate->frame` may change if the shadow stack is materialized
between when the new frame is allocated and when its set as
`tstate->frame`. The window of time is very small, but it's
possible. We must ensure that if `tstate->frame` changes, the newly
allocated frame's `f_back` is updated to point to it. Otherwise, we can
end up with a missing frame on the Python stack.

I've fixed it in this diff by updating the frame's `f_back` to `tstate->frame`
immediately before the frame is linked in the interpreter.

The following scenario illustrates how this can happen and is codified in a
unit test.

T0:

At time T0, the stacks look like:

```
Shadow Stack        PyFrame Stack
------------        ------------
f0                  f0
f1
```

- `f0` is interpreted and has called `f1`.
- `f1` is jit-compiled and is running in shadow-frame mode. The stack
  hasn't been materialized, so there is no entry for `f1` on the
  pyframe stack.

T1:

At time T1, `f1` calls `f2`. `f2` is interpreted and has a variable keyword
parameter (**kwargs). The call to `f2` enters `PyEval_EvalCodeWithName`.
`PyEval_EvalCodeWithName` allocates a new `PyFrameObject`, `p`, to run
`f2`. At this point the stacks still look the same, however, notice
that `p->f_back` points to `f0`, not `f1`.

```
Shadow Stack        PyFrame Stack
------------        ------------
f0                  f0              <--- p->f_back points to f0
f1
```

T2:

At time T2, `PyEval_EvalCodeWithName` has allocated the `PyFrameObject`,
`p`, for `f2`, and allocates a new dictionary for the variable keyword
parameter. The dictionary allocation triggers GC. During GC an object
is collected with a finalizer that materializes the stack. (The most
common way for this to happen is through an unawaited coroutine. The
coroutine's finalizer will call `_PyErr_WarnUnawaitedCoroutine` which
materializes the stack.)

Notice that the stacks match after materialization, however,
`p->f_back` still points to `f0`.

```
Shadow Stack        PyFrame Stack
------------        ------------
f0                  f0              <--- p->f_back points to f0
f1                  f1
```

T3:

At time T3, control has transferred from `PyEval_EvalCodeWithName`
to the interpreter loop. The interpreter loop has set `tstate->frame`
to the frame it was passed, `p`. Now the stacks are mismatched, as
we've erased f1 from the pyframe stack.

```
Shadow Stack        PyFrame Stack
------------        ------------
f0                  f0
f1                  f2
f2
```

T4:

At time T4, `f2` finishes execution and returns into `f1`.

```
Shadow Stack        PyFrame Stack
------------        ------------
f0                  f0
f1
```

T5:

At time T5, `f1` finishes execution and attempts to return. Since the
stack was materialized, it expects to find a `PyFrameObject` for
`f1` at the top of the PyFrame stack and aborts when it does not.

Reviewed By: swtaarrs

Differential Revision: D32661319

fbshipit-source-id: 1ec0274082