Implement ContextManager _no_profile_inner in C
Summary:
This provides a C accelerator for the inner methods in the ContextDecorator class.  This would be a lot more complicated if we didn't already have a wrapper for awaitables in the class loader, but because we do it's somewhat straight forward.  But regardless it's async, so it's somwhat complicated.

In order to remain as compatible as possible w/ existing code we keep the existing `_no_profile_inner` function around but replace it's vectorcall with a C implementation.

For sync methods it's pretty straight forward - we just need to do what the underlying code does.  This adds caches based upon the types version tag for lookups of `_recreate_cm`, `__enter__`, and `__exit__`. And we invoke the function between `__enter__` and `__exit__`.  For the async version we just need to return the async awaitable and then invoke `__exit__` once it completes.  But we also need to handle the case where we're not eagerly evaluated, in which case `__enter__` should be called on the first step of the function.

This adds a weird state transition of the awaitable wrapper where we start off with a step before invoking the iterable, and then we transition the state it holds from the context decorator wrapper to the exit function which we usually hold onto.

The existing awaitable wrapper is made to be generic so that we can pass a generic callback and state object into it - and it also supports a callback to run before running the iterable.

In order to support keeping the existing function around we need fast access back to our context manager wrapper state.  To achieve this the wrapper is actually a weak ref object, which we assign into the functions `func_weakreflist` - this then lets us usually get to it as the very first item in the weak ref list.  This is basically identical to do what we do in shadow code, and like in shadow code, we have some interesting ref counting tricks.  Because the weak ref is only references by the function we setup a callback on it so that we can dec ref the context manager wrapper when the function goes away.

On a simple microbenchmark with a nop context decorator, ran with the JIT w/ shadow frames, this seems to be about 4x faster.

Reviewed By: sinancepel

Differential Revision: D32407250

fbshipit-source-id: dce968febb