Add funcTypeChecks pass
Summary:
Add a pass that goes through every instruction and verifies that each
instruction's operands are of an appropriate type.

This serves as a good sanity check right now that the JIT doesn't do
anything unexpected with instructions that require specific types in
their operands, and will serve as a safety check once we add
profile-guided type specialization and introduce potential type widening
into the JIT.

This diff also modifies `Constraint::kMatchOther` to ensure that we can
appropriately type check instructions that take operands of the "same"
type (whose join type subclasses a provided type) - such as an instance of
`IntBinaryOp` passing type-checking when given `CInt8` and `Bottom`.

By adding `funcTypeCheck` to the compiler pass, running a canary of
distillery exposes a few wrongly typed opcodes, so this diff also fixes those
types.

Reviewed By: swtaarrs

Differential Revision: D32022280

fbshipit-source-id: 879b451bf9