Fix issue w/ async function returning tuple
Summary:
There's some tests which are failing because they return a tuple, and when we go and set that as the `StopIteration` value we're losing the values past the first one.

It turns out there's a helper for this that we should be using, so this just switches over to using the helper.

Reviewed By: carljm

Differential Revision: D32820069

fbshipit-source-id: 3344a9dee5