Fix: Allow materialization of runing generator frames that aren't on the stack
Summary:
A generator will be marked as running but will not be on the stack when
it appears as a predecessor in a chain of generators into which an
exception was thrown. For example, given an "await stack" of
coroutines like the following, where ` a <- b` indicates a `a` awaits `b`,

```
coro0 <- coro1 <- coro2
```

if someone does `coro0.throw(...)`, then `coro0` and `coro1` will be
marked as running but will not appear on the stack while `coro2` is
handling the exception.

Prior to this change, we would trigger an assert and crash if, while
handling an exception that was thrown into it, a generator accessed a
predecessor's frame  (e.g. if `coro2` accessed `coro1.cr_frame`).

Reviewed By: swtaarrs

Differential Revision: D32870469

fbshipit-source-id: 93ef262c81