dev guide initial skeleton
Summary:
Initial draft/skeleton of a JIT dev guide, mostly motivated by having somewhere to record
the rationale and restrictions of JIT preloading.

The aim here was to give a high-level architectural overview, pointers in the right direction
for further code reading for each step, and some key invariants to keep in mind when
working on the JIT code. There's of course a lot more that could go here, this is just a start;
many of the sections are basically just stubs.

Reviewed By: swtaarrs

Differential Revision: D32836619

fbshipit-source-id: 1b259b2105