Fix Github issue 59, don't allow generic types w/ unknown parmaeters
Summary:
Currently we allow indexing generic types with an unknown type.  For CheckedDict this results us in constructing the type with:

`INVOKE_FUNCTION CheckedDict[..., object](CheckedDict[..., unknown_type]).__new__`

Which results in an exception because we're invoking `__new__` and passing a different/incompatible type.

This changes it so `CheckedDict[...].__new__` no longer exists, and instead we make sure the tp_alloc slot does all of the necessary initialization.  And then for construction of the dict we now use `TP_ALLOC`.

This isn't really a general purpose solution because if we had a generic immutable type we would still need to have ` __new__` which took arguments, in which case we'd hit this again.  But we also can't simply disallow unknown types here because we already have uses of unknown types like `CheckedDict[..., Set[...]]` which are perfectly valid and treated as object - we just happen to construct them with literals which don't have this problem.

There's probably a couple of possible solutions for the immutable type problem. One might be to add a way to invoke __new__ that takes the type descriptor and passes the resolved type in as the 1st argument (i.e. a TP_NEW opcode). What might be better would be adding an opcode which turns a type descriptor into the underlying type. Then we could emit the __new__ call with the compiler known type instead of emitting the various loads and binary subscrs required to produce the type. That might not require any specialization of the __new__ handling - we could just use that to emit any generic type references.

Reviewed By: sinancepel

Differential Revision: D32842269

fbshipit-source-id: a8bfdb348c