backport bpo-42202 to cinder
Summary: bpo-42202 avoids creating a dictionary for annotations until the annotations are actually accessed, instead storing the annotations in a tuple with the name being the even index and value being the odd.  Because we don't do a lot of reflection on annotations this should be a decent memory win for us.  Further because we frequently use `from __future__ import annotations` this should also have some marginal startup gains as we just get to do a simple `LOAD_CONST` to produce the entire annotations tuple.

Reviewed By: wmeehan

Differential Revision: D32632149

fbshipit-source-id: 801da2bda5