Stop patching return addresses for JIT-compiled functions with materialized frames
Summary:
Our current strategy for unlinking materialized shadow frames is to
patch the return address of the function so that it returns into a
trampoline that unlinks the Python frame object. That works and is
efficient, however, the calling function will appear as the frame
unlink trampoline in profiling tools.

This moves to a model where we always check whether we need to unlink
a Python frame object in the epilogue. This appears to be ~perf neutral
and doesn't affect profiling tools.

Reviewed By: swtaarrs

Differential Revision: D33053258

fbshipit-source-id: b9ab6b09b3