Clear prev when popping shadow frames
Summary:
We weren't clearing the `prev` pointer for generator shadow frames when
they suspended. If the `get_entire_call_stack_as_qualnames` function
encounters a generator without an awaiter set it falls back to walking the
sync stack. This is to handle the case where an eagerly executed coroutine
awaits a non-eagerly executed coroutine. However, if the coroutine suspends,
then prev would point to an address on the stack that is no longer valid.

Instead of fixing this only for generators I've updated the pop
function to always clear prev in order to prevent similar situations
in the future.

Reviewed By: tekknolagi

Differential Revision: D33161976

fbshipit-source-id: 8db284e52b