start emitting `__final_method_names__` as a slot for static classes
Summary:
Submitting this diff on its own to get feedback on how we want to handle the final slots with regards to subclassing. I chose a simple approach where we populate a `__final_method_slots__` by looking through the entire MRO and folding all final methods together, which gives us all method names in a tuple.

Won't land this until we also support final method invokes, but just want to ease review.

The open questions I have for the implementation are:

1. Should we be doing something smarter where we look for the `__final_method_slots__` of the base and extending that during codegen? I can see this being more efficient, but might be a bit fussy.
2. What data structure should I use for these slots to support faster final override checks? This is obviously introducing O(n^2) scaling (at the very least, we'll need to turn things into a more easily searchable structure in `type_new` with this data structure). I can see us wanting to keep things as a sorted tuple and binary search, or simply use a set, but curious whether there's precedent of what other parts of the runtime do the solve this problem.
3. Should we be caching the `get_own_final_method_slots()` function's results? I thought about doing so, and feared that it might unsafe, but maybe it's not and I can just get away with it :P

Reviewed By: carljm

Differential Revision: D32956316

fbshipit-source-id: c685c1e376